requirements.txt
asgiref==3.7.2
bcrypt==4.1.3
Django==4.2.11
django-axes==8.0.0
django-cors-headers==3.14.0
django-crispy-forms==2.4
django-dbbackup==5.0.0
django-defender==0.9.8
django-environ==0.12.0
django-filter==23.5
django-googledrive-storage @ git+https://github.com/torre76/django-googledrive-storage.git@2f5aac22236b08a63f234bd5ee8c0e9db3c5e3d5
django-phonenumber-field==7.1.0
django-rest-swagger==2.2.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
GDAL @ file:///C:/GDAL-3.8.4-cp312-cp312-win_amd64.whl
phonenumbers==8.13.34
pillow==10.3.0
psycopg2-binary==2.9.10
pymemcache==4.0.0
pytz==2024.1
redis==6.4.0
sqlparse==0.4.4
tzdata==2025.2
